refactor(drive-integration): remove google-docs-agent-improvements flag
