--- apps/drive-integration/src/locations/Page/components/mainpage/MainPageView.tsx
@@ -1,91 +1,0 @@
-import { ArrowRightIcon } from '@contentful/f36-icons';
-import {
-  Button,
-  Card,
-  Flex,
-  Grid,
-  Heading,
-  Layout,
-  Note,
-  Paragraph,
-} from '@contentful/f36-components';
-import tokens from '@contentful/f36-tokens';
-import { OAuthConnector } from './OAuthConnector';
-
-interface MainPageViewProps {
-  oauthToken: string;
-  isOAuthConnected: boolean;
-  isOAuthLoading: boolean;
-  isOAuthBusy: boolean;
-  onConnectGoogleDrive: () => Promise<void>;
-  onDisconnectGoogleDrive: () => Promise<void>;
-  onSelectFile: () => void;
-}
-
-export const MainPageView = ({
-  oauthToken,
-  isOAuthConnected,
-  isOAuthLoading,
-  isOAuthBusy,
-  onConnectGoogleDrive,
-  onDisconnectGoogleDrive,
-  onSelectFile,
-}: MainPageViewProps) => {
-  return (
-    <Layout.Body>
-      <Flex
-        flexDirection="column"
-        gap="spacingXl"
-        style={{ maxWidth: '900px', margin: `${tokens.spacingL} auto` }}>
-        <Flex justifyContent="space-between" alignItems="center">
-          <Heading marginBottom="none">Drive Integration</Heading>
-          <OAuthConnector
-            isOAuthConnected={isOAuthConnected}
-            isOAuthBusy={isOAuthBusy}
-            onConnect={onConnectGoogleDrive}
-            onDisconnect={onDisconnectGoogleDrive}
-          />
-        </Flex>
-        <Card padding="large">
-          <Flex
-            flexDirection="column"
-            alignItems="flex-start"
-            justifyContent="space-between"
-            gap="spacingS"
-            style={{ width: '100%' }}>
-            <Heading marginBottom="spacingS">Select your file</Heading>
-            <Grid columns="4fr 1fr" style={{ width: '100%' }}>
-              <Grid.Item>
-                <Paragraph>
-                  Create entries using existing content types from a Drive file.
-                  <br />
-                  Get started by selecting the .doc file you would like to use. Only Google Doc
-                  files are supported. Sheets, Slides, and PDFs will not appear in the file picker.
-                </Paragraph>
-              </Grid.Item>
-              <Grid.Item justifySelf="end">
-                <Button
-                  variant="primary"
-                  size="small"
-                  isDisabled={!oauthToken}
-                  onClick={onSelectFile}
-                  endIcon={<ArrowRightIcon />}>
-                  Select
-                </Button>
-              </Grid.Item>
-            </Grid>
-            <Note variant="neutral">
-              This app only creates new entries. Any existing entries that need to be referenced
-              must be assigned after the draft entries are created.
-            </Note>
-            {!isOAuthLoading && !isOAuthConnected && (
-              <Note variant="warning" style={{ width: '100%' }}>
-                Please connect to Drive before selecting your file.
-              </Note>
-            )}
-          </Flex>
-        </Card>
-      </Flex>
-    </Layout.Body>
-  );
-};
