+++ apps/drive-integration/src/utils/constants/agent.ts
@@ -3,10 +3,8 @@ export const WORKFLOW_AGENT_ID = 'google-docs-workflow-agent';
 export const POLL_INTERVAL_MS = 10000; // 10 seconds
 
 const MAX_POLL_TIME_MS = 20 * 60 * 1000; // 20 minutes
-const EXTENDED_POLL_TIME_MS = 60 * 60 * 1000; // 60 minutes (google-docs-agent-improvements flag)
 
 export const MAX_POLL_ATTEMPTS = Math.floor(MAX_POLL_TIME_MS / POLL_INTERVAL_MS);
-export const EXTENDED_MAX_POLL_ATTEMPTS = Math.floor(EXTENDED_POLL_TIME_MS / POLL_INTERVAL_MS);
 
 // Agents-api writes PENDING_REVIEW status before the suspendPayload metadata flushes.
 // Allow this many consecutive PENDING_REVIEW polls with no suspendPayload before giving up.
